Remove explicit normalisation of numerator Blatt-Weisskopf factors
Summary:
As per T93, the explicit normalisation of the Blatt-Weisskopf factors that multiply the amplitude appears to be unnecessary and has the undesired consequence of requiring p_0 or p*_0 to be calculable.
This, in turn, requires the use of the ad-hoc effective mass for resonances with a pole mass above the "upper threshold" value of parent mass - bachelor mass.

These changes remove the explicit normalisation of these factors, while retaining it for the resonance factor that enters the mass-dependent width.

At the same time, have introduced the possibility to use the ad-hoc effective mass for resonances below threshold in LauGounarisSakurai (that feature was previously only available in LauRelBreitWigner).

Test Plan:
Have generated toy MC events using GenFit3pi and fitted them back with code in current master branch and with these changes.

The values of the integrals change, as expected, but the generated events are identical and the fit parameters show differences only at the 6th significant digit.

Reviewers: jback, mwhitehe, dpohanlon

Reviewed By: jback, mwhitehe, dpohanlon

Tags: #laura

Differential Revision: https://phab.hepforge.org/D29