Allow increase of C++ standard via environment variable
Summary:
Tweak to CMake configuration to allow the version of the C++ standard to be
modified via the environment variable `CMAKE_CXX_STANDARD`.
The implementation only allows the change if the specified standard is newer
than our default, currently C++17.

Test Plan:
Configured the project with the environment variable set to 14 and 20 and
obtained the expected behaviour: no change, standard bumped to C++20.

Reviewers: jback

Reviewed By: jback

Tags: #laura

Differential Revision: https://phab.hepforge.org/D105